fix too late initialization of pa_module.userdata
m->userdata needs to be set before any "goto fail" statements, because
otherwise there may have been some initialization done that that
doesn't get appropriately cleaned up.

This issue manifested itself in a crash when I had not set
DBUS_SESSION_BUS_ADDRESS correctly when running PulseAudio. That made
the initialization of module-murphy-ivi fail, and because cleanup was
not done in pa__done(), there was a stale sink unlink hook slot left
unfreed. The hook slot pointed to an unavailable callback function,
which caused a crash when a sink was unlinked.

Change-Id: I50a9aaa000ad441a8ad9db9a5822e62b5aa0b099